added update fish functions

+++ .dotfiles/tag-smu/config/fish/functions/update.fish
@@ -1,23 +1,19 @@
-function update --description "Updates Linux apps, brew, npm and their installed packages"
-    sudo apt update
-    sudo apt upgrade -y
-    sudo apt dist-upgrade -y
-    sudo apt full-upgrade -y
-    sudo apt autoremove -y
-    sudo apt clean
- 
+function update --description "Updates Linux/MacOS apps, brew and their installed packages"
+    if uname = "Linux"
+        sudo apt update
+        sudo apt upgrade -y
+        sudo apt dist-upgrade -y
+        sudo apt full-upgrade -y
+        sudo apt autoremove -y
+        sudo apt clean
+     end
+
+     if uname = "Darwin"
+        sudo softwareupdate --install --all
+     end
+
      if type -q brew
-         brew update
-         brew upgrade
-         brew cleanup
-     end
- 
-     if type -q npm
-         sudo npm install npm@latest -g
- 
-         if type -q npx
-             npx npm-check --global --update-all
-         end
-     end
+        brew-update
+    end
  end
